+++ src/collective/cover/layout.py
@@ -159,10 +159,13 @@ class TileList(grok.View):
     def get_tile_metadata(self, tile_name):
         tile_type = getUtility(ITileType, tile_name)
         tile = self.context.restrictedTraverse(str(tile_name))
+        title = tile.short_name
+        if not title:
+            title = tile_type.title
         tile_metadata = {
             'icon': tile_type.icon,
             'description': tile_type.description,
-            'title': tile_type.title,
+            'title': title,
             'is_configurable': tile.is_configurable and 1 or 0
         }