+++ src/collective/cover/tiles/carousel.py
@@ -37,6 +37,7 @@ class CarouselTile(ListTile):
     index = ViewPageTemplateFile("templates/carousel.pt")
     is_configurable = True
     is_editable = True
+    short_name = _(u"msg_short_name_carousel", default=u"Carousel")
 
     def populate_with_object(self, obj):
         super(CarouselTile, self).populate_with_object(obj)  # check permission